requirements.txt
bleach
django
inflection
isodate
pyparsing
rdflib
requests
django-bootstrap-pagination
python-dateutil
elasticsearch
-e git+https://github.com/mesemus/autobreadcrumbs.git#egg=autobreadcrumbs
cachetools
django-model-utils